fix(homepage): correct totalKm unit from meters to km
f1api.dev returns circuitLength as `"<N>km"` where the digits are
**meters** (e.g. Albert Park `"5278km"` = 5.278 km), not kilometers.
The previous mapper digit-stripped the value and treated the result
as km, producing a season total roughly 1000x too large — caught by
the persona critique before stats-aware users hit it on a "real data
only" surface.

- `Season.totalKm` is now `Double` (was `Int`) to preserve the
  per-circuit precision across a 24-race sum.
- `GetSeasonUseCase.toSeason()` does `digits / 1000.0` per race.
- `HomepageScreen.kt` section 2 renders `%.1f` (the label "Total km
  covered" was added in the prior clarify pass so no unit is
  repeated in the value).
- KDoc rewritten in both model and mapper to call out the meter
  wire format so the next person to touch the contract doesn't
  re-introduce the bug.

+++ app/src/main/java/com/anpurnama/f1_app/f1/model/Season.kt
@@ -7,7 +7,15 @@ data class Season(
     val year: Int,
     val races: List<Race>,
     val completedGp: Int,
-    val totalKm: Int,
+    /**
+     * Total **kilometers** of track distance across completed rounds
+     * (sum of `circuitLength / 1000` for races with a winner). Stored
+     * as `Double` because the wire value is in meters (e.g. `"5412km"`
+     * for Bahrain = 5.412 km) and a 24-race season can carry three
+     * decimal places across the sum. Renders as `"%.1f"` to drop the
+     * trailing noise.
+     */
+    val totalKm: Double,
     val totalLaps: Int,
     val progressPercent: Float,
 )
